+++ src/com/yanncebron/m68kplugin/editor/M68kCommaFlipper.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 The Authors
+ * Copyright 2026 The Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,6 +18,10 @@ package com.yanncebron.m68kplugin.editor;
 
 import com.intellij.openapi.editor.actions.FlipCommaIntention;
 import com.intellij.psi.PsiElement;
+import com.yanncebron.m68kplugin.lang.psi.M68kAdm;
+import com.yanncebron.m68kplugin.lang.psi.M68kInstruction;
+import com.yanncebron.m68kplugin.lang.psi.M68kPsiElement;
+import com.yanncebron.m68kplugin.lang.psi.M68kPsiTreeUtil;
 import com.yanncebron.m68kplugin.lang.psi.expression.M68kExpression;
 import org.jetbrains.annotations.NotNull;
 
@@ -30,7 +34,16 @@ final class M68kCommaFlipper implements FlipCommaIntention.Flipper {
 
   @Override
   public boolean canFlip(@NotNull PsiElement left, @NotNull PsiElement right) {
-    return left instanceof M68kExpression &&
-      right instanceof M68kExpression;
+    if (left instanceof M68kExpression && right instanceof M68kExpression) {
+      return true;
+    }
+
+    if (left instanceof M68kAdm && right instanceof M68kAdm) {
+      M68kPsiElement instructionOrDirective = M68kPsiTreeUtil.getContainingInstructionOrDirective(left);
+      return instructionOrDirective instanceof M68kInstruction m68kInstruction &&
+        M68kMoveLeftRightHandler.getMovableSubElementsForInstruction(m68kInstruction).length == 2;
+    }
+
+    return false;
   }
 }